+++ bindings/nodejs/CHANGELOG.md
@@ -25,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Account::{burn(), consolidateOutputs(), createAliasOutput(), meltNativeToken(), mintNativeToken(), createNativeToken(), mintNfts(), sendTransaction(), sendNativeTokens(), sendNft()}` methods;
 - `Client::outputIds()` method;
-- `UnlockableByAddress` QueryParameter;
+- `GenericQueryParameter, UnlockableByAddress` types;
 
 ## 1.0.11 - 2023-09-14